requirements.txt
paho-mqtt>=1.6.1
requests>=2.31.0
opencv-python>=4.8.0
pillow>=10.0.0
sqlalchemy>=2.0.0
python-dateutil>=2.8.2
pyyaml>=6.0
flask>=3.0.0
